requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
fastapi==0.116.2
idna==3.10
logging-loki==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.38
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
